+++ CHANGELOG.md
@@ -6,9 +6,58 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [5.0.1]
+## [5.0.6]
 ### Added
-- [Docker documentation](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/README.md#L108)
+- `cascade={"persist", "remove"}` to [`User::$recoveryCodes`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Entity/User.php#L33)
+- Continuous integration update
+
+## [5.0.5] - 2018-10-09
+### Changed
+- [`README.md`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/README.md)
+- Text of import database process in [Composer setup script](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Composer/Setup.php#L152)
+
+### Fixed
+- `doctrine/orm` version at `php >= 5.5.0` in [`composer.json`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/composer.json#L24)
+
+## [5.0.4] - 2018-10-08
+### Added
+- Dynamic generated [`public_path`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/config/settings.php#L7)
+- [Docker condition](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Controller/UserController.php#L77) to reCAPTCHA validation
+- `remove-cache` [Composer script](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/composer.json#L52)
+
+### Changed
+- [`README.md`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/README.md)
+- [`GeneralUtility::encryptPassword`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Utility/GeneralUtility.php#L15) from `mcrypt_create_iv` to `random_bytes`
+- Return of [`GeneralUtility::getUserIP`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Utility/GeneralUtility.php#L124)
+
+### Removed
+- `public_path` from [`additional-settings.dist.php`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/config/additional-settings.dist.php)
+
+### Fixed
+- `public_path` setup in [Composer setup script](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Composer/Setup.php#L203)
+
+## [5.0.3] - 2018-10-02
+### Fixed
+- Typo in [Composer setup script](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Composer/Setup.php)
+- ErrorException `mysqli::__construct(): (HY000/2002): No such file or directory` in [Composer setup script](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Composer/Setup.php)
+
+## [5.0.2] - 2018-09-30
+### Added
+- [Composer callbacks](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/composer.json#L55)
+- [Composer setup script](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Composer/Setup.php)
+- [Database port](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/config/settings.php#L57)
+- [`scripts.js`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/public/js/scripts.js)
+
+### Changed
+- [Docker documentation](https://github.com/InsanityMeetsHH/Slim-Skeleton#installation-with-docker)
+- `docker-dump.sql` is now [`db-dump.sql`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/sql/db-dump.sql)
+
+### Removed
+- `"start": "php -S localhost:8080 -t public",` from [`composer.json`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/composer.json) (use docker instead)
+
+## [5.0.1] - 2018-09-23
+### Added
+- [Docker documentation](https://github.com/InsanityMeetsHH/Slim-Skeleton#installation-with-docker)
 - [`Dockerfile`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/Dockerfile)
 - [`docker-dump.sql`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/sql/docker-dump.sql)
 
@@ -20,7 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `default_domain` from `locale` in [`config/settings.php`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/config/settings.php#L32) and [`config/additional-settings.dist.php`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/config/additional-settings.dist.php#L23)
 - `default_domain` from [`LanguageExtension`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/src/Twig/LanguageExtension.php#L79)
 
-## [5.0.0] - 2018-08-17
+## [5.0.0] - 2018-09-17
 ### Added
 - Localized website by domain (exampl.de / de.example.com)
 - Localized website by session (exampl.com for every language)
@@ -41,7 +90,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSS](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/public/css/styles.css#L64) for Google reCAPTCHA widget
 - Registration form
 - Localized flash messages for registration form
-- Documentation in README.md
+- Documentation in [`README.md`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/README.md)
 - `{{ glc }}` in [`templates/layouts/layout.html.twig`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/templates/layouts/layout.html.twig#L4)
 - Website session script in [`templates/layouts/layout.html.twig`](https://github.com/InsanityMeetsHH/Slim-Skeleton/blob/master/templates/layouts/layout.html.twig#L33)
 - `public/favicon.ico`